refactor authentication func to service
