+++ src/pages/PlantManager.js
@@ -2,28 +2,13 @@ import axios from "axios";
 import React, { Fragment, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import NaviBar from "../components/NaviBar/NaviBar";
+import { checkAdminPermission } from "../service/Authentication";
 
 const PlantManager = () => {
-
-  const navigate = useNavigate()
-
+  const navigate = useNavigate();
   useEffect(() => {
-    const token = localStorage.getItem('token');
-    axios.get("http://localhost:8000/getpermission", {
-      "headers":
-      {'Authorization': "Bearer " + token}
-    }).then((response) => {
-      console.log(response.response)
-    }).catch((error) => {
-      console.log(error.response.data.detail);
-      if (error.response.data.detail == "Signature has expired"){
-        localStorage.removeItem('token');
-        localStorage.removeItem('name');
-      }
-      navigate('/shelf')
-    })
-
-  }, [])
+    checkAdminPermission(navigate);
+  }, []);
 
   return (
     <Fragment>
